requirements.txt
tensorflow==2.14.0
--extra-index-url https://download.pytorch.org/whl/cu116
torch==2.0.1
torchvision==0.15.2
fastai==2.7.12
fastapi==0.108.0
imageio==2.31.1
keras==2.14.0
matplotlib==3.7.1
numpy==1.23.5
opencv_contrib_python==4.8.0.74
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
pandas==2.0.2
Pillow==10.2.0
pydantic==1.10.9
rembg==2.0.50
scikit_learn==1.2.2
seaborn==0.13.1
timm==0.9.2
tqdm==4.65.0
vit_pytorch==1.2.4
uvicorn==0.22.0
